// src/lib/types.ts
export interface TaskDue {
    date: string
}

export interface RawTask {
    id: string
    content: string
    checked: boolean
    completed_at: string | null
    due: TaskDue | null
    project_id: string | null
    labels: string[]
    child_order: number
    is_deleted?: boolean
}

export interface EnrichedTask extends RawTask {
    project_name: string
    label_names: string[]
    due_date: Date | null
    has_time: boolean
}
// ...
export interface TaskBackendConfig {
    apiToken?: string
}